requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
dependency-injector==4.41.0
Django==5.0.3
django-admin-rangefilter==0.12.1
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
django-rest-knox==4.2.0
django-storages==1.14.2
djangorestframework==3.15.1
drf-standardized-errors==0.13.0
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.21
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
